refactor: series auto install mark
