+++ packages/vchart/src/chart/index.ts
@@ -1,23 +1,134 @@
-export * from './line';
-export * from './area';
-export * from './bar';
-export * from './histogram';
-export * from './pie';
-export * from './rose';
-export * from './map';
-export * from './funnel';
-export * from './radar';
-export * from './word-cloud';
-export * from './sequence';
-export * from './progress/circular';
-export * from './progress/linear';
-export * from './range-column';
-export * from './scatter';
-export * from './sunburst';
-export * from './circle-packing';
-export * from './box-plot';
-export * from './gauge';
-export * from './heatmap';
+/* eslint-disable no-duplicate-imports */
+/**
+ * @description export all chart modules
+ */
+import type { IAreaChartSpec } from './area';
+import { AreaChart } from './area';
+import type { IBarChartSpec, IBar3dChartSpec } from './bar';
+import { BarChart, Bar3dChart } from './bar';
+import type { IBoxPlotChartSpec } from './box-plot';
+import { BoxPlotChart } from './box-plot';
+import type { ICirclePackingChartSpec } from './circle-packing';
+import { CirclePackingChart } from './circle-packing';
+import type { ICommonChartSpec } from './common';
+import { CommonChart } from './common';
+import type { IFunnelChartSpec, IFunnel3dChartSpec } from './funnel';
+import { FunnelChart, Funnel3dChart } from './funnel';
+import type { IGaugeChartSpec } from './gauge';
+import { GaugeChart } from './gauge';
+import type { IHeatmapChartSpec } from './heatmap';
+import { HeatmapChart } from './heatmap';
+import type { IHistogramChartSpec, IHistogram3dChartSpec } from './histogram';
+import { HistogramChart, Histogram3dChart } from './histogram';
+import type { ILineChartSpec } from './line';
+import { LineChart } from './line';
+import type { IMapChartSpec } from './map';
+import { MapChart } from './map';
+import type { IPieChartSpec, IPie3dChartSpec } from './pie';
+import { PieChart, Pie3dChart } from './pie';
+import type { ICircularProgressChartSpec } from './progress/circular';
+import { CircularProgressChart } from './progress/circular';
+import type { ILinearProgressChartSpec } from './progress/linear';
+import { LinearProgressChart } from './progress/linear';
+import type { IRadarChartSpec } from './radar';
+import { RadarChart } from './radar';
+import type { IRangeAreaChartSpec } from './range-area';
+import { RangeAreaChart } from './range-area';
+import type { IRangeColumnChartSpec, IRangeColumn3dChartSpec } from './range-column';
+import { RangeColumnChart, RangeColumn3dChart } from './range-column';
+import type { IRoseChartSpec } from './rose';
+import { RoseChart } from './rose';
+import type { ISankeyChartSpec } from './sankey';
+import { SankeyChart } from './sankey';
+import type { IScatterChartSpec } from './scatter';
+import { ScatterChart } from './scatter';
+import type { ISequenceChartSpec } from './sequence';
+import { SequenceChart } from './sequence';
+import type { ISunburstChartSpec } from './sunburst';
+import { SunburstChart } from './sunburst';
+import type { ITreemapChartSpec } from './treemap';
+import { TreeMapChart } from './treemap';
+import type { IWaterfallChartSpec } from './waterfall';
+import { WaterfallChart } from './waterfall';
+import type { IWordCloudChartSpec, IWordCloud3dChartSpec } from './word-cloud';
+import { WordCloudChart, WordCloud3dChart } from './word-cloud';
+import { BaseChart } from './base-chart';
+import type { ICartesianChartSpec } from './cartesian';
+import { CartesianChart } from './cartesian';
+import type { IPolarChartSpec } from './polar';
+import { PolarChart, RoseLikeChart, ProgressLikeChart } from './polar';
 
-export * from './common';
-export * from './base-chart';
+export {
+  AreaChart,
+  BarChart,
+  Bar3dChart,
+  BoxPlotChart,
+  CirclePackingChart,
+  CommonChart,
+  FunnelChart,
+  Funnel3dChart,
+  GaugeChart,
+  HeatmapChart,
+  HistogramChart,
+  Histogram3dChart,
+  LineChart,
+  MapChart,
+  PieChart,
+  Pie3dChart,
+  CircularProgressChart,
+  LinearProgressChart,
+  RadarChart,
+  RangeColumnChart,
+  RangeColumn3dChart,
+  RangeAreaChart,
+  RoseChart,
+  SankeyChart,
+  ScatterChart,
+  SunburstChart,
+  SequenceChart,
+  WaterfallChart,
+  WordCloudChart,
+  WordCloud3dChart,
+  TreeMapChart,
+  BaseChart,
+  PolarChart,
+  ProgressLikeChart,
+  RoseLikeChart,
+  CartesianChart
+};
+
+export type {
+  IAreaChartSpec,
+  IBarChartSpec,
+  IBar3dChartSpec,
+  IBoxPlotChartSpec,
+  ICirclePackingChartSpec,
+  ICommonChartSpec,
+  IFunnelChartSpec,
+  IFunnel3dChartSpec,
+  IGaugeChartSpec,
+  IHeatmapChartSpec,
+  IHistogramChartSpec,
+  IHistogram3dChartSpec,
+  ILineChartSpec,
+  IMapChartSpec,
+  IPieChartSpec,
+  IPie3dChartSpec,
+  ICircularProgressChartSpec,
+  ILinearProgressChartSpec,
+  IRadarChartSpec,
+  IRangeColumnChartSpec,
+  IRangeColumn3dChartSpec,
+  IRangeAreaChartSpec,
+  IRoseChartSpec,
+  IScatterChartSpec,
+  ISankeyChartSpec,
+  ISequenceChartSpec,
+  ISunburstChartSpec,
+  ITreemapChartSpec,
+  IWaterfallChartSpec,
+  IWordCloud3dChartSpec,
+  IWordCloudChartSpec,
+  IPolarChartSpec,
+  ICartesianChartSpec
+};
