+++ packages/vchart/src/component/index.ts
@@ -1,1 +1,101 @@
-export * from './axis';
+/* eslint-disable no-duplicate-imports */
+/**
+ * @description export all component modules
+ */
+import type { IScrollBarSpec } from './data-zoom/scroll-bar';
+import { ScrollBar } from './data-zoom/scroll-bar';
+import type { IDataZoomSpec } from './data-zoom/data-zoom';
+import { DataZoom } from './data-zoom/data-zoom';
+import { CustomMark } from './custom-mark';
+import type { IBrushSpec } from './brush';
+import { Brush } from './brush';
+import type { IMapLabelSpec } from './map-label';
+import { MapLabelComponent } from './map-label';
+import type {
+  ICartesianLinearAxisSpec,
+  ICartesianBandAxisSpec,
+  ICartesianTimeAxisSpec,
+  ICartesianAxisSpec
+} from './axis/cartesian/index';
+import { CartesianAxis, CartesianLinearAxis, CartesianBandAxis, CartesianTimeAxis } from './axis/cartesian/index';
+import type { IPolarAxisSpec, IPolarBandAxisSpec, IPolarLinearAxisSpec } from './axis/polar';
+import { PolarAxis, PolarBandAxis, PolarLinearAxis } from './axis/polar';
+import type { IDiscreteLegendSpec } from './legend/discrete';
+import { DiscreteLegend } from './legend/discrete';
+import type { IContinuousLegendSpec } from './legend/continuous';
+import { ContinuousLegend } from './legend/continuous';
+import type { IIndicatorSpec } from './indicator';
+import { Indicator } from './indicator';
+import type { ITitleSpec } from './title';
+import { Title } from './title';
+import type { IGeoCoordinateSpec } from './geo';
+import { GeoCoordinate } from './geo';
+import type { ICartesianCrosshairSpec, IPolarCrosshairSpec } from './crosshair';
+import { CartesianCrossHair, PolarCrossHair } from './crosshair';
+import type { IPlayerSpec } from './player';
+import { Player } from './player';
+import type { IMarkLineSpec } from './marker/mark-line';
+import { MarkLine } from './marker/mark-line';
+import type { IMarkAreaSpec } from './marker/mark-area';
+import { MarkArea } from './marker/mark-area';
+import type { IMarkPointSpec } from './marker/mark-point';
+import { MarkPoint } from './marker/mark-point';
+import type { ITooltipSpec } from './tooltip';
+import { Tooltip } from './tooltip';
+import type { ILabelSpec } from './label';
+import { Label } from './label';
+
+export {
+  ScrollBar,
+  DataZoom,
+  CustomMark,
+  Brush,
+  MapLabelComponent,
+  CartesianAxis,
+  CartesianBandAxis,
+  CartesianLinearAxis,
+  CartesianTimeAxis,
+  PolarAxis,
+  PolarBandAxis,
+  PolarLinearAxis,
+  DiscreteLegend,
+  ContinuousLegend,
+  Indicator,
+  Title,
+  GeoCoordinate,
+  CartesianCrossHair,
+  PolarCrossHair,
+  Player,
+  MarkArea,
+  MarkLine,
+  MarkPoint,
+  Tooltip,
+  Label
+};
+
+export type {
+  IScrollBarSpec,
+  IBrushSpec,
+  ICartesianAxisSpec,
+  ICartesianBandAxisSpec,
+  ICartesianCrosshairSpec,
+  ICartesianLinearAxisSpec,
+  ICartesianTimeAxisSpec,
+  IContinuousLegendSpec,
+  IDataZoomSpec,
+  IDiscreteLegendSpec,
+  IGeoCoordinateSpec,
+  IIndicatorSpec,
+  ILabelSpec,
+  IMapLabelSpec,
+  IMarkAreaSpec,
+  IMarkLineSpec,
+  IMarkPointSpec,
+  IPlayerSpec,
+  IPolarAxisSpec,
+  IPolarBandAxisSpec,
+  IPolarCrosshairSpec,
+  IPolarLinearAxisSpec,
+  ITitleSpec,
+  ITooltipSpec
+};
