--- packages/vchart/src/export/layout.ts
@@ -1,12 +1,0 @@
-/* eslint-disable no-duplicate-imports */
-/**
- * @description export all layout modules
- */
-import { Layout } from '../layout/index';
-import type { IGridLayoutSpec } from '../layout/grid-layout/grid-layout';
-import { GridLayout } from '../layout/grid-layout/grid-layout';
-import { Layout3d } from '../layout/layout3d';
-import type { ILayoutSpec } from '../layout/interface';
-
-export { Layout, GridLayout, Layout3d };
-export type { ILayoutSpec, IGridLayoutSpec };
