+++ packages/vchart/src/index.ts
@@ -1,14 +1,14 @@
 import { VChart } from './vchart-all';
 export default VChart;
 
-export * from './export/core';
+export * from './core';
 
 // chart
-export * from './export/chart';
-export * from './export/series';
-export * from './export/mark';
-export * from './export/component';
-export * from './export/layout';
+export * from './chart';
+export * from './series';
+export * from './mark';
+export * from './component';
+export * from './layout';
 /**
  * spec
  */
