+++ packages/vchart/src/series/index.ts
@@ -1,0 +1,145 @@
+/* eslint-disable no-duplicate-imports */
+/**
+ * @description export all mark modules
+ */
+import { WaterfallSeries } from './waterfall/waterfall';
+import type { IWaterfallSeriesSpec } from './waterfall/interface';
+import { BoxPlotSeries } from './box-plot/box-plot';
+import type { IBoxPlotSeriesSpec } from './box-plot/interface';
+import { LineSeries } from './line/line';
+import type { ILineSeriesSpec } from './line/interface';
+import { BarSeries } from './bar/bar';
+import { Bar3dSeries } from './bar/bar-3d';
+import type { IBar3dSeriesSpec, IBarSeriesSpec } from './bar/interface';
+import { RangeColumnSeries } from './range-column/rangeColumn';
+import { RangeColumn3dSeries } from './range-column/rangeColumn-3d';
+import type { IRangeColumn3dSeriesSpec, IRangeColumnSeriesSpec } from './range-column/interface';
+import { RangeAreaSeries } from './range-area/range-area';
+import type { IRangeAreaSeriesSpec } from './range-area/interface';
+import { MapSeries } from './map/map';
+import type { IMapSeriesSpec } from './map/interface';
+import { PieSeries } from './pie/pie';
+import { Pie3dSeries } from './pie/pie-3d';
+import type { IPie3dSeriesSpec } from './pie/interface';
+import type { IPieSeriesSpec } from './pie/interface';
+import { ScatterSeries } from './scatter/scatter';
+import type { IScatterSeriesSpec } from './scatter/interface';
+import { RoseSeries } from './rose/rose';
+import type { IRoseSeriesSpec } from './rose/interface';
+import { RadarSeries } from './radar/radar';
+import type { IRadarSeriesSpec } from './radar/interface';
+import { AreaSeries } from './area/area';
+import type { IAreaSeriesSpec } from './area/interface';
+import { DotSeries } from './dot/dot';
+import type { IDotSeriesSpec } from './dot/interface';
+import { LinkSeries } from './link/link';
+import type { ILinkSeriesSpec } from './link/interface';
+import { CircularProgressSeries } from './progress/circular/circular';
+import type { ICircularProgressSeriesSpec } from './progress/circular/interface';
+import { LinearProgressSeries } from './progress/linear/linear';
+import type { ILinearProgressSeriesSpec } from './progress/linear/interface';
+import { WordCloudSeries } from './word-cloud/word-cloud';
+import { WordCloud3dSeries } from './word-cloud/word-cloud-3d';
+import type { IWordCloud3dSeriesSpec, IWordCloudSeriesSpec } from './word-cloud/interface';
+import { FunnelSeries } from './funnel/funnel';
+import { Funnel3dSeries } from './funnel/funnel-3d';
+import type { IFunnel3dSeriesSpec, IFunnelSeriesSpec } from './funnel/interface';
+import { SunburstSeries } from './sunburst/sunburst';
+import type { ISunburstSeriesSpec } from './sunburst/interface';
+import { CirclePackingSeries } from './circle-packing/circle-packing';
+import type { ICirclePackingSeriesSpec } from './circle-packing/interface';
+import { SankeySeries } from './sankey/sankey';
+import type { ISankeySeriesSpec } from './sankey/interface';
+import { TreeMapSeries } from './treemap/treemap';
+import type { ITreemapSeriesSpec } from './treemap/interface';
+import type { IGaugePointerSeriesSpec, IGaugeSeriesSpec } from './gauge';
+import { GaugePointerSeries, GaugeSeries } from './gauge';
+import { HeatmapSeries } from './heatmap/heatmap';
+import type { IHeatmapSeriesSpec } from './heatmap/interface';
+import { BaseSeries } from './base/base-series';
+import type { ICartesianSeriesSpec } from './cartesian';
+import { CartesianSeries } from './cartesian';
+import { PolarSeries } from './polar/polar';
+import type { IPolarSeriesSpec } from './polar/interface';
+import type { IProgressLikeSeriesSpec } from './polar/progress-like';
+import { ProgressLikeSeries } from './polar/progress-like';
+import type { IRoseLikeSeriesSpec } from './polar/rose-like';
+import { RoseLikeSeries } from './polar/rose-like';
+
+import type { ISeries } from './interface';
+
+export {
+  WaterfallSeries,
+  Bar3dSeries,
+  BarSeries,
+  BoxPlotSeries,
+  LineSeries,
+  RadarSeries,
+  RangeAreaSeries,
+  RangeColumn3dSeries,
+  RangeColumnSeries,
+  MapSeries,
+  Pie3dSeries,
+  PieSeries,
+  SankeySeries,
+  ScatterSeries,
+  SunburstSeries,
+  RoseLikeSeries,
+  RoseSeries,
+  AreaSeries,
+  DotSeries,
+  LinearProgressSeries,
+  LinkSeries,
+  CirclePackingSeries,
+  CircularProgressSeries,
+  WordCloud3dSeries,
+  WordCloudSeries,
+  Funnel3dSeries,
+  FunnelSeries,
+  TreeMapSeries,
+  GaugePointerSeries,
+  GaugeSeries,
+  HeatmapSeries,
+  BaseSeries,
+  CartesianSeries,
+  PolarSeries,
+  ProgressLikeSeries
+};
+
+export type {
+  ISeries,
+  IRoseLikeSeriesSpec,
+  IAreaSeriesSpec,
+  IBar3dSeriesSpec,
+  IBarSeriesSpec,
+  IBoxPlotSeriesSpec,
+  ICartesianSeriesSpec,
+  ICirclePackingSeriesSpec,
+  ICircularProgressSeriesSpec,
+  IDotSeriesSpec,
+  IFunnel3dSeriesSpec,
+  IFunnelSeriesSpec,
+  IGaugePointerSeriesSpec,
+  IGaugeSeriesSpec,
+  IHeatmapSeriesSpec,
+  ILineSeriesSpec,
+  ILinearProgressSeriesSpec,
+  ILinkSeriesSpec,
+  IMapSeriesSpec,
+  IPie3dSeriesSpec,
+  IPieSeriesSpec,
+  IPolarSeriesSpec,
+  IProgressLikeSeriesSpec,
+  IRadarSeriesSpec,
+  IRangeAreaSeriesSpec,
+  IRangeColumn3dSeriesSpec,
+  IRangeColumnSeriesSpec,
+  IRoseSeriesSpec,
+  ISankeySeriesSpec,
+  IScatterSeriesSpec,
+  ISunburstSeriesSpec,
+  ITreemapSeriesSpec,
+  IWaterfallSeriesSpec,
+  IWordCloud3dSeriesSpec,
+  IWordCloudSeriesSpec
+};
