+++ packages/vchart/src/vchart-simple.ts
@@ -1,10 +1,10 @@
 /**
  * @description 包含基础的折柱饼图，提供坐标轴、离散图例以及 tooltip、crosshair、label 组件
  */
-import { VChart } from './export/core';
-import { LineChart, BarChart, PieChart } from './export/chart';
-import { LineSeries, BarSeries, PieSeries } from './export/series';
-import { SymbolMark, LineMark, RectMark, ArcMark, TextMark, PathMark } from './export/mark';
+import { VChart } from './core';
+import { LineChart, BarChart, PieChart } from './chart';
+import { LineSeries, BarSeries, PieSeries } from './series';
+import { SymbolMark, LineMark, RectMark, ArcMark, TextMark, PathMark } from './mark';
 
 import {
   CartesianLinearAxis,
@@ -14,7 +14,7 @@ import {
   Tooltip,
   CartesianCrossHair,
   Label
-} from './export/component';
+} from './component';
 
 // charts
 VChart.useChart([LineChart, BarChart, PieChart]);
@@ -22,8 +22,8 @@ VChart.useChart([LineChart, BarChart, PieChart]);
 // series
 VChart.useSeries([LineSeries, BarSeries, PieSeries]);
 
-// marks
-VChart.useMark([SymbolMark, LineMark, RectMark, ArcMark, TextMark, PathMark]);
+// // marks
+// VChart.useMark([SymbolMark, LineMark, RectMark, ArcMark, TextMark, PathMark]);
 
 // components
 VChart.useComponent([
